Profile page

+++ frontend/src/pages/ProfilePage.vue
@@ -1,8 +1,13 @@
 <template>
-  <div>
+  <div class="profile-page">
     <h2 class="title">{{ user.username }}</h2>
-    <div >
+    <h3>{{user.email}}</h3>
+    <div v-if="isAdmin">
       <h3 @click="deleteUser">DELETE</h3>
+    </div>
+    <div v-if="isAdmin">
+      <h3 v-if="!profileUserIsModerator" @click="makeModerator">Make Moderator</h3>
+      <h3 v-if="profileUserIsModerator" @click="removeModerator">Remove Moderator</h3>
     </div>
   </div>
 </template>
@@ -10,16 +15,57 @@
 <script>
 export default {
   props: ['user'],
+  computed: {
+    isAdmin() {
+      return this.$store.getters.currentUser.userRole === "admin";
+    },
+    profileUserIsModerator() {
+      return this.user.userRole === "moderator";
+    }
+  },
   methods: {
     async deleteUser() {
+
+      if (this.$store.getters.currentUser.userRole !== "admin") { return }
       console.log("DELETING USER");
 
-      let res = await fetch(`http://localhost:3000/rest/deleteusers/${this.user.id}`, {
+      let res = await fetch(`/rest/v1/deleteusers/${this.user.id}`, {
         method: 'DELETE'
       });
       res = await res.json();
       console.log(res);
-
+      this.goToHomePage();
+    },
+    logoutUser() {
+      // NEEDED or only ADMIN CAN DELETE USER?
+    },
+    goToHomePage() {
+        this.$router.push("/");
+    },
+    async makeModerator() {
+      
+      let res = await fetch(`/rest/v1/user/${this.user.id}`, {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body:
+          JSON.stringify({ "userRole": "moderator" })
+      });
+      res = await res.json();
+      console.log(res);
+    },
+    async removeModerator() {
+      let res = await fetch(`/rest/v1/user/${this.user.id}`, {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body:
+          JSON.stringify({ "userRole": "member" })
+      });
+      res = await res.json();
+      console.log(res);
     }
   },
 
@@ -28,8 +74,14 @@ export default {
 </script>
 
 <style scoped>
+.profile-page {
+  background-color: grey;
+  height: 80vh;
+}
 
 .title {
+  padding-top: 40px;
+  margin: 0px;
   text-align: center;
   text-decoration: underline;
 }
